Text fixes and redo of machine translations

+++ lib/djvAV/Render2D.cpp
@@ -533,8 +533,8 @@ namespace djv
                 {
                     auto logSystem = context->getSystemT<LogSystem>();
                     std::stringstream ss;
-                    ss << "Texture atlas count: " << _textureAtlasCount << "\n";
-                    ss << "Texture atlas size: " << _textureAtlasSize;
+                    ss << "Texture atlas count: " << static_cast<size_t>(_textureAtlasCount) << "\n";
+                    ss << "Texture atlas size: " << static_cast<size_t>(_textureAtlasSize);
                     logSystem->log("djv::AV::Render::Render2D", ss.str());
                 }
                 p.textureAtlas.reset(new TextureAtlas(
